backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
chromadb==0.5.20
google-genai>=1.0.0
httpx==0.27.2
python-dotenv==1.0.1
lxml==5.3.0
requests==2.32.3
pydantic==2.9.2
